FE-98: Rename `useDeleteFinanceRecordMutation` to `useDeleteFinanceRecord`
